Create a new set of bindings to use for serializing the work graph nodes to the instant execution cache.
Also allow a codec to create a nested context with a new isolate and codec to use for serializing nested objects. Use this to maintain a single isolate for all work nodes with nested isolates for each task instance.
