+++ subprojects/instant-execution/src/test/kotlin/org/gradle/instantexecution/serialization/codecs/BeanCodecTest.kt
@@ -99,14 +99,14 @@ class BeanCodecTest {
 
     private
     inline fun <R> MutableIsolateContext.withIsolateMock(block: () -> R): R =
-        withIsolate(IsolateOwner.OwnerGradle(mock())) {
+        withIsolate(IsolateOwner.OwnerGradle(mock()), codecs().userTypesCodec) {
             block()
         }
 
     private
     fun writeContextFor(encoder: Encoder) =
         DefaultWriteContext(
-            encodings = codecs().userTypesCodec,
+            codec = codecs().userTypesCodec,
             encoder = encoder,
             logger = mock(),
             problemHandler = mock()
@@ -115,7 +115,7 @@ class BeanCodecTest {
     private
     fun readContextFor(inputStream: ByteArrayInputStream) =
         DefaultReadContext(
-            decoding = codecs().userTypesCodec,
+            codec = codecs().userTypesCodec,
             decoder = KryoBackedDecoder(inputStream),
             logger = mock(),
             beanPropertyReaderFactory = BeanPropertyReader.factoryFor(mock())
@@ -127,7 +127,9 @@ class BeanCodecTest {
         fileCollectionFactory = mock(),
         fileResolver = mock(),
         instantiator = mock(),
-        listenerManager = mock()
+        listenerManager = mock(),
+        projectStateRegistry = mock(),
+        taskNodeFactory = mock()
     )
 
     @Test
